Added the prepared statement and confirmation for delete

+++ mvc/app/view/blog/index.php
@@ -52,7 +52,7 @@ session_start();
                         $date = $row['Date_Time'];
                         echo "<p><b>".$row['Post']."</b></p><p>".$row['Date_Time']."</p>";
                         echo "<button><a href='edit/".$row['Id']."'>Edit</a></button>";
-                        echo "<button><a href='/blog/delete/".$row['Id']."'>Delete</a></button>";
+                        echo "<button><a onclick='return confirm(\"Are you sure you want to delete the post?\")' href='/blog/delete/".$row['Id']."'>Delete</a></button>";
                         if($row['Edit_Time'] != NULL)
                         {
                             echo "(Edited: ".$row['Edit_Time'].")<br>";
